Converted code
My initial plan to use an automated tool to convert from java to
javascript does not work. While the resulting java does probably run, it
can not be used as base for future development. Partly because comments
are stripped way, but also because the code feels a a quite unnatural
for javascript.  So I do everything by hand, which is both boring and
error-prone.